+++ app/Http/Controllers/DepartamentosController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Http\Controllers;
 
 use App\Models\Departamento;
@@ -10,15 +9,17 @@ class DepartamentosController extends Controller
 {
     public function __construct()
     {
-        if (Auth::check() && !Auth::user()->can('departamento.show')) {
-            abort(403, 'Acción no autorizada !');
-        }
+
     }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        if (Auth::check() && ! Auth::user()->can('departamento.show')) {
+            abort(403, 'Acción no autorizada !');
+        }
+
         $departamentos = Departamento::all();
         return view('backend.departamentos.index', compact('departamentos'));
     }
@@ -36,26 +37,30 @@ class DepartamentosController extends Controller
      */
     public function store(Request $request)
     {
+        if (Auth::check() && ! Auth::user()->can('departamento.create')) {
+            abort(403, 'Acción no autorizada !');
+        }
+
         $request->validate([
-            'nombre' => 'required',
+            'nombre'      => 'required',
             'descripcion' => 'nullable',
-            'hora_ini' => 'required',
-            'hora_fin' => 'required',
-            'latitud' => 'nullable',
-            'longitud' => 'nullable',
+            'hora_ini'    => 'required',
+            'hora_fin'    => 'required',
+            'latitud'     => 'nullable',
+            'longitud'    => 'nullable',
         ], [
-            'nombre.required' => 'El nombre es obligatorio',
+            'nombre.required'   => 'El nombre es obligatorio',
             'hora_ini.required' => 'La hora de entrada es obligatoria',
             'hora_fin.required' => 'La hora de salida es obligatoria',
         ]);
 
-        $departamento = new Departamento();
-        $departamento->nombre = $request->nombre;
+        $departamento              = new Departamento();
+        $departamento->nombre      = $request->nombre;
         $departamento->descripcion = $request->descripcion;
-        $departamento->hora_ini = $request->hora_ini;
-        $departamento->hora_fin = $request->hora_fin;
-        $departamento->latitud = $request->latitud;
-        $departamento->longitud = $request->longitud;
+        $departamento->hora_ini    = $request->hora_ini;
+        $departamento->hora_fin    = $request->hora_fin;
+        $departamento->latitud     = $request->latitud;
+        $departamento->longitud    = $request->longitud;
         $departamento->save();
 
         session()->flash('message', 'Departamento agregado correctamente');
@@ -88,26 +93,30 @@ class DepartamentosController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (Auth::check() && ! Auth::user()->can('departamento.edit')) {
+            abort(403, 'Acción no autorizada !');
+        }
+
         $request->validate([
-            'nombre' => 'required',
+            'nombre'      => 'required',
             'descripcion' => 'nullable',
-            'hora_ini' => 'required',
-            'hora_fin' => 'required',
-            'latitud' => 'nullable',
-            'longitud' => 'nullable',
+            'hora_ini'    => 'required',
+            'hora_fin'    => 'required',
+            'latitud'     => 'nullable',
+            'longitud'    => 'nullable',
         ], [
-            'nombre.required' => 'El nombre es obligatorio',
+            'nombre.required'   => 'El nombre es obligatorio',
             'hora_ini.required' => 'La hora de entrada es obligatoria',
             'hora_fin.required' => 'La hora de salida es obligatoria',
         ]);
 
-        $departamento = Departamento::findOrFail($id);
-        $departamento->nombre = $request->nombre;
+        $departamento              = Departamento::findOrFail($id);
+        $departamento->nombre      = $request->nombre;
         $departamento->descripcion = $request->descripcion;
-        $departamento->hora_ini = $request->hora_ini;
-        $departamento->hora_fin = $request->hora_fin;
-        $departamento->latitud = $request->latitud;
-        $departamento->longitud = $request->longitud;
+        $departamento->hora_ini    = $request->hora_ini;
+        $departamento->hora_fin    = $request->hora_fin;
+        $departamento->latitud     = $request->latitud;
+        $departamento->longitud    = $request->longitud;
         $departamento->save();
 
         session()->flash('message', 'Departamento actualizado correctamente');
@@ -124,6 +133,10 @@ class DepartamentosController extends Controller
      */
     public function destroy(Request $request, string $id)
     {
+        if (Auth::check() && ! Auth::user()->can('departamento.delete')) {
+            abort(403, 'Acción no autorizada !');
+        }
+
         $departamento = Departamento::findOrFail($id);
         $departamento->delete();
